# add codegen for serialization of hamt types