--- a/go.mod
+++ b/go.mod
@@ -4,7 +4,9 @@ require (
 	github.com/ipfs/go-block-format v0.0.2
 	github.com/ipfs/go-cid v0.0.3
 	github.com/ipfs/go-ipld-cbor v0.0.3
+	github.com/multiformats/go-multihash v0.0.1
 	github.com/polydawn/refmt v0.0.0-20190221155625-df39d6c2d992
 	github.com/spaolacci/murmur3 v0.0.0-20180118202830-f09979ecbc72
-	github.com/whyrusleeping/cbor-gen v0.0.0-20190822002707-4e02357de5c1
+	github.com/whyrusleeping/cbor-gen v0.0.0-20190822231004-8db835b09a5a
+	golang.org/x/xerrors v0.0.0-20190717185122-a985d3407aa7
 )