This is mostly good, theres a bit of funk that i'm trying to figure out how to address, but thats mostly changes in the codegen itself, and not here (notably, i don't always want big.Int to turn into a tagged cbor bignum)