@whyrusleeping 2019-08-22
I need to write some benchmarks, but just running the tests shows some nice preliminary results:

before:
```
start flush
flush took:  193ns
start flush
flush took:  29.36µs
Total size is: 7740515, size of keys+vals: 6600000, overhead: 1.17
&{4713 100000 map[1:41259 2:16451 3:8613]}
start flush
flush took:  492.862011ms
finds took:  386.000758ms
thingy1 bafyreicmfb6plrpg6s3a3nllfa5t5kg3zpvxp525n235mvfg2ogs73dbda
&{ 0 bafyreicmfb6plrpg6s3a3nllfa5t5kg3zpvxp525n235mvfg2ogs73dbda}
PASS
ok  	github.com/ipfs/go-hamt-ipld	2.298s
```

After:
```
start flush
flush took:  573ns
start flush
flush took:  2.323µs
Total size is: 7724621, size of keys+vals: 6600000, overhead: 1.17
&{4694 100000 map[1:41151 2:16475 3:8633]}
start flush
flush took:  50.8159ms
finds took:  232.45173ms
thingy1 bafy2bzacedwfefsr5nposuqzr76ox2tkg23lifkopqed5sd3t34ke7vlczmea
&{ 0 bafy2bzacedwfefsr5nposuqzr76ox2tkg23lifkopqed5sd3t34ke7vlczmea}
PASS
ok  	github.com/ipfs/go-hamt-ipld	1.164s
```